Fit-Me-Up/AI/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.2
flatbuffers==25.9.23
fonttools==4.60.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.26.0
idna==3.11
jax==0.6.2
jaxlib==0.6.2
jiter==0.11.0
kiwisolver==1.4.9
matplotlib==3.10.6
mediapipe==0.10.14
ml_dtypes==0.5.3
numpy==2.1.1
openai==2.8.1
opencv-contrib-python==4.12.0.88
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==25.0
pillow==10.4.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scipy==1.15.3
serpapi==0.1.5
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.40.0
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
